Add Google Places Autocomplete to the Shelter Form
This also adds a couple new columns for state, zip, and a google place
ID for use with other google products possibly.

These fields are all filled in automatically when the user accepts a
Google Places autocomplete suggestion.

Wiring up the javascript was a bit harrowing, since Rails is not the
most opinionated about how to structure javascript.